+++ locals-naming.tf
@@ -1,0 +1,8 @@
+locals {
+  # Naming locals/constants
+  name_prefix = lower(var.name_prefix)
+  name_suffix = lower(var.name_suffix)
+
+  public_ip_name   = coalesce(var.public_ip_custom_name, data.azurecaf_name.pip_nat.result)
+  nat_gateway_name = coalesce(var.nat_gateway_custom_name, data.azurecaf_name.nat.result)
+}