+++ r-nat-gateway.tf
@@ -8,7 +8,7 @@ resource "azurerm_public_ip" "pip" {
   sku                 = "Standard"
 
   domain_name_label = var.public_ip_domain_name_label
-  reverse_fqdn = var.public_ip_reverse_fqdn
+  reverse_fqdn      = var.public_ip_reverse_fqdn
 
   tags = merge(
     local.default_tags,