+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Breaking
   * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
 
 # v4.1.0 - 2022-04-15