+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Unreleased
 
 Added
-  * AZ-XXX : Add public ip Domain dns label option
-  * AZ-XXX : Add public ip Reverse Fqdn option
+  * AZ-917 : Add public ip Domain dns label option
+  * AZ-917 : Add public ip Reverse Fqdn option
 
 # v4.1.0 - 2022-04-15