+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Unreleased
 
+<<<<<<< HEAD
 Added
   * AZ-917 : Add public ip Domain dns label option
   * AZ-917 : Add public ip Reverse Fqdn option
+=======
+Breaking
+  * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
+>>>>>>> master
 
 # v4.1.0 - 2022-04-15